Refactor team and agency health performance dashboards to use the same calculations
